AIRBOTS/requirement.txt
# Core Robotics & Autopilot
mavsdk==1.4.4
pymavlink==2.4.37

# Computer Vision & AI
ultralytics==8.3.172
torch==1.13.1
torchvision==0.14.1
# opencv-python-headless==4.8.0.76
opencv-python==4.8.0.76        # for gui
pycairo
PyGObject

# Hardware Control (Jetson Nano only; comment out if not needed)
jetson-gpio==1.0.0

# Audio & UI
pygame==2.5.0

# Utilities
numpy==1.23.5
pyyaml==6.0.1
python-dateutil==2.8.2
tqdm==4.66.1

# ROS Noetic Integration (Optional, for advanced users)
rospkg==1.5.0
catkin-pkg==0.5.2
# The following ROS message packages are typically installed via apt, not pip:
# sudo apt install ros-noetic-mavros-msgs ros-noetic-sensor-msgs

# Optional (Documentation)
sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Testing
pytest==7.4.2
coverage       # <---- Optional, test coverage reports
pytest-mock    # <---- Optional, for mocking in pytest

# No changes to versions unless a vulnerability arises.
